requirements.txt
# requirements.txt
# 核心依赖
python-dotenv==1.0.0
supabase==1.1.1
yfinance==0.2.28
pandas==2.0.3

# 可选依赖（如果使用其他API）
requests==2.31.0

# 开发依赖（可选）
pytest==7.4.0
black==23.7.0
flake8==6.0.0